+++ Tests/test_registration.py
@@ -1,7 +1,8 @@
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from locators import AllLocators
-from data import name, incorrect_password, new_email, new_password
+from data import name, incorrect_password
+from helpers import new_email, new_password
 
 class TestRegistration:
     def test_successful_registration(self, browser):
@@ -12,6 +13,12 @@ class TestRegistration:
         browser.find_element(*AllLocators.REGISTER_EMAIL).send_keys(new_email)
         browser.find_element(*AllLocators.LOGIN_PASSWORD).send_keys(new_password)
         browser.find_element(*AllLocators.REGISTER_BUTTON).click()
+
+        success_element = WebDriverWait(browser, 10).until(
+            EC.presence_of_element_located(AllLocators.LOGIN_BUTTON_SIGNIN),
+            message="Кнопка входа не появилась после регистрации"
+        )
+        assert success_element.is_displayed(), "Регистрация не удалась: кнопка входа не отображается"
 
     def test_failed_registration(self, browser):
         browser.find_element(*AllLocators.ACCOUNT_BUTTON).click()
