+++ locators.py
@@ -5,14 +5,16 @@ class AllLocators:
     ACCOUNT_BUTTON = (By.XPATH, '//a[@href="/account"]')  # Личный кабинет
     LOGIN_BUTTON = (By.XPATH, '//button[contains(@class, "button_button_size_large") and text()="Войти в аккаунт"]') # Войти в аккаунт
 
-    #Конструктор
-    BUNS_SECTION = (By.XPATH, '//*[contains(@class, "text_type_main-default") and text()="Булки"]') #раздел Булки
-    SAUCE_SECTION = (By.XPATH, '//*[contains(@class, "text_type_main-default") and text()="Соусы"]')# раздел Соусы
-    TOPPINGS_SECTION = (By.XPATH, '//*[contains(@class, "text_type_main-default") and text()="Начинки"]')  #раздел Начинки
-    MAKE_ORDER = (By.XPATH, '//button[contains(text(),"Оформить заказ")]') #Оформить заказ
-    ACTIVE_SECTION = (By.XPATH, '//*[@id="root"]/div/main/section[1]/div[1]') #блок с навигацией
+    #Конструктор (табы)
+    BUNS_SECTION = (By.XPATH, '//div[contains(@class, "tab") and .="Булки"]') #раздел Булки
+    SAUCE_SECTION = (By.XPATH, '//div[contains(@class, "tab") and .="Соусы"]')# раздел Соусы
+    TOPPINGS_SECTION = (By.XPATH, '//div[contains(@class, "tab") and .="Начинки"]')  #раздел Начинки
 
-#Логин
+    #Активные табы
+    MAKE_ORDER = (By.XPATH, '//button[contains(text(),"Оформить заказ")]') #Оформить заказ
+    ACTIVE_SECTION = (By.XPATH, '//div[contains(@class, "tab_tab_type_current")]') #блок с навигацией
+
+    #Логин
     LOGIN_EMAIL = (By.XPATH, "//input [@name= 'name']") #Email
     LOGIN_PASSWORD = (By.XPATH, "//input[@type='password' and @name='Пароль']") #Пароль, можно использовать в Регистрации
     LOGIN_BUTTON_SIGNIN = (By.XPATH, "//button[contains(text(), 'Войти')]") #Войти
@@ -24,8 +26,8 @@ class AllLocators:
     HEADER_LOGIN = (By.XPATH, "//h2[contains (text),'Вход')]") #Вход (заголовок)
 
     #Регистрация
-    REGISTER_NAME = (By.XPATH, "//fieldset[1]//input")  # Имя
-    REGISTER_EMAIL = (By.XPATH, "//fieldset[2]//input")  # Email
+    REGISTER_NAME = (By.XPATH, "//label[text()='Имя']/following-sibling::input")  # Имя
+    REGISTER_EMAIL = (By.XPATH, "//label[text()='Email']/following-sibling::input")  # Email
     REGISTER_BUTTON = (By.XPATH, "//button[text()='Зарегистрироваться']")  # Зарегистрироваться
     ERROR_PASSWORD = (By.XPATH, "//p[contains(@class, 'input__error') and contains(text(), 'Некорректный пароль')]")  # Некорректный пароль
     SIGNIN_LINK = (By.XPATH, "//a[contains (text(), 'Войти')]")  # Войти (можно использовать на странице забыли пароль)
